+++ src/API/Delegates.php
@@ -11,16 +11,14 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace ArkEcosystem\Client\API\One;
-
-use ArkEcosystem\Client\API\AbstractAPI;
+namespace ArkEcosystem\Client\API;
 
 /**
- * This is the transactions resource class.
+ * This is the delegates resource class.
  *
  * @author Brian Faust <brian@ark.io>
  */
-class Transactions extends AbstractAPI
+class Delegates extends AbstractAPI
 {
     /**
      * Get all accounts.
@@ -31,11 +29,11 @@ class Transactions extends AbstractAPI
      */
     public function all(array $query = []): array
     {
-        return $this->get('transactions', $query);
+        return $this->get('delegates', $query);
     }
 
     /**
-     * Get a transaction by the given id.
+     * Get a block by the given id.
      *
      * @param string $id
      *
@@ -43,30 +41,32 @@ class Transactions extends AbstractAPI
      */
     public function show(string $id): array
     {
-        return $this->get('transactions/get', compact('id'));
+        return $this->get("delegates/{$id}");
     }
 
     /**
-     * Get all unconfirmed transactions.
-     *
-     * @param array $query
-     *
-     * @return array
-     */
-    public function allUnconfirmed(array $query = []): array
-    {
-        return $this->get('transactions/unconfirmed', $query);
-    }
-
-    /**
-     * Get an unconfirmed transaction by the given id.
+     * Get all blocks for the given delegate.
      *
      * @param string $id
+     * @param array  $query
      *
      * @return array
      */
-    public function showUnconfirmed(string $id): array
+    public function blocks(string $id, array $query = []): array
     {
-        return $this->get('transactions/unconfirmed/get', compact('id'));
+        return $this->get("delegates/{$id}/blocks", $query);
+    }
+
+    /**
+     * Get all voters for the given delegate.
+     *
+     * @param string $id
+     * @param array  $query
+     *
+     * @return array
+     */
+    public function voters(string $id, array $query = []): array
+    {
+        return $this->get("delegates/{$id}/voters", $query);
     }
 }
