--- tests/API/Two/PeersTest.php
@@ -1,49 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of Ark PHP Client.
- *
- * (c) Ark Ecosystem <info@ark.io>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace ArkEcosystem\Tests\Client\API\Two;
-
-use ArkEcosystem\Tests\Client\TestCase;
-
-/**
- * This is the peers resource test class.
- *
- * @author Brian Faust <brian@ark.io>
- * @covers \ArkEcosystem\Client\API\Two\Peers
- */
-class PeersTest extends TestCase
-{
-    /** @test */
-    public function all_calls_correct_url()
-    {
-        $this->assertResponse(2, 'GET', 'peers', function ($mock) {
-            return $mock->all();
-        });
-    }
-
-    /** @test */
-    public function show_calls_correct_url()
-    {
-        $this->assertResponse(2, 'GET', 'peers/dummy', function ($mock) {
-            return $mock->show('dummy');
-        });
-    }
-
-    /**
-     * @return string
-     */
-    protected function getApiClass()
-    {
-        return \ArkEcosystem\Client\API\Two\Peers::class;
-    }
-}
